pyproject.toml
[project]
name = "eu5-mod-orchestrator"
version = "0.1.0"
description = "Reusable orchestration engine for Europa Universalis V mod build pipelines."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
    "typer>=0.20.0",
]

[project.scripts]
eu5-orchestrator = "eu5_mod_orchestrator.cli:app"

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
